remove old sobel
